Address review comments. Refactor out AnthemCommunicationManager
