+++ bundles/org.openhab.binding.anthem/src/main/resources/OH-INF/i18n/anthem.properties
@@ -1,50 +1,10 @@
-# add-on
-
-addon.anthem.name = Anthem Binding
-addon.anthem.description = This is the binding for Anthem AV preamp/processors.
-
-# thing types
-
-thing-type.anthem.label = anthem
-thing-type.anthem.description = Thing for Anthem AV processor
-
-# thing types config
-
-thing-type.config.anthem.host.label = Network Address
-thing-type.config.anthem.host.description = Host name or IP address of the Anthem AV processor
-thing-type.config.anthem.port.label = Network Port
-thing-type.config.anthem.port.description = Network port of the Anthem AV processor
-thing-type.config.anthem.reconnectintervalminutes.label = Reconnect Interval
-thing-type.config.anthem.reconnectintervalminutes.description = The time to wait between reconnection attempts (in minutes)
-thing-type.config.anthem.comanddelaymsec.label = Command Delay
-thing-type.config.anthem.comanddelaymsec.description = The delay between commands sent to the processor (in milliseconds)
-
-# channel group types
-
-channel-group-type.anthem.zone1.label = Main Zone
-channel-group-type.anthem.zone1.description = Controls zone 1 (the main zone) of the processor
-channel-group-type.anthem.zone2.label = Zone 2
-channel-group-type.anthem.zone2.description = Controls zone 2 of the processor
-
-# channel types
-
-channel-type.anthem.volumedb.label = Volume dB
-channel-type.anthem.volumedb.description = Set the volume level dB between -90 and 0
-channel-type.anthem.mutetoggle.label = Mute Toggle
-channel-type.anthem.mutetoggle.description = Toggle the mute between ON and OFF
-channel-type.anthem.activeinput.label = Active Input
-channel-type.anthem.activeinput.description = Selects the active input source
-channel-type.anthem.activeinputshortname.label = Active Input Short Name
-channel-type.anthem.activeinputshortname.description = Short friendly name of the active input source
-channel-type.anthem.activeinputlongname.label = Active Input Long Name
-channel-type.anthem.activeinputlongname.description = Long friendly name of the active input source
-
+# Thing status detail messages
 thing-status-detail-initializing = Initializing
 thing-status-detail-connecting = Connecting
+thing-status-detail-reconnecting = Reconnecting
 thing-status-detail-unknownhost = Unknown host
 thing-status-detail-invalidport = Invalid port number
 thing-status-detail-openerror = Error opening Anthem connection. Check log
 thing-status-detail-interrupted = Interrupted
 thing-status-detail-ioerror = I/O Error
-thing-status-detail-runtimeexception = Runtime exception
 thing-status-detail-invalidconfig = Invalid Anthem thing configuration
