+++ bundles/org.openhab.binding.anthem/src/main/resources/OH-INF/thing/thing-types.xml
@@ -5,45 +5,45 @@
 	xsi:schemaLocation="https://openhab.org/schemas/thing-description/v1.0.0 https://openhab.org/schemas/thing-description-1.0.0.xsd">
 
 	<thing-type id="anthem">
-		<label>@text/thing-type.anthem.label</label>
-		<description>@text/thing-type.anthem.description</description>
+		<label>anthem</label>
+		<description>Thing for Anthem AV processor</description>
 
 		<channel-groups>
 			<channel-group id="1" typeId="zone">
-				<label>@text/channel-group-type.anthem.zone1.label</label>
-				<description>@text/channel-group-type.anthem.zone1.description</description>
+				<label>Main Zone</label>
+				<description>Controls zone 1 (the main zone) of the processor</description>
 			</channel-group>
 
 			<channel-group id="2" typeId="zone">
-				<label>@text/channel-group-type.anthem.zone2.label</label>
-				<description>@text/channel-group-type.anthem.zone2.description</description>
+				<label>Zone 2</label>
+				<description>Controls zone 2 of the processor</description>
 			</channel-group>
 		</channel-groups>
 
 		<config-description>
 			<parameter name="host" type="text" required="true">
-				<label>@text/thing-type.config.anthem.host.label</label>
-				<description>@text/thing-type.config.anthem.host.description</description>
+				<label>Network Address</label>
+				<description>Host name or IP address of the Anthem AV processor</description>
 				<context>network-address</context>
 			</parameter>
 
 			<parameter name="port" type="integer" required="true">
-				<label>@text/thing-type.config.anthem.port.label</label>
-				<description>@text/thing-type.config.anthem.port.description</description>
+				<label>Network Port</label>
+				<description>Network port number of the Anthem AV processor</description>
 				<default>14999</default>
 				<advanced>true</advanced>
 			</parameter>
 
 			<parameter name="reconnectIntervalMinutes" type="integer" required="true">
-				<label>@text/thing-type.config.anthem.reconnectintervalminutes.label</label>
-				<description>@text/thing-type.config.anthem.reconnectintervalminutes.description</description>
+				<label>Reconnect Interval</label>
+				<description>The time to wait between reconnection attempts (in minutes)</description>
 				<default>2</default>
 				<advanced>true</advanced>
 			</parameter>
 
 			<parameter name="comandDelayMsec" type="integer" required="true">
-				<label>@text/thing-type.config.anthem.comanddelaymsec.label</label>
-				<description>@text/thing-type.config.anthem.comanddelaymsec.description</description>
+				<label>Command Delay</label>
+				<description>The delay between commands sent to the processor (in milliseconds)</description>
 				<default>100</default>
 				<advanced>true</advanced>
 			</parameter>
@@ -58,7 +58,6 @@
 			<channel id="volume" typeId="system.volume"/>
 			<channel id="volumeDB" typeId="volumeDB"/>
 			<channel id="mute" typeId="system.mute"/>
-			<channel id="muteToggle" typeId="muteToggle"/>
 			<channel id="activeInput" typeId="activeInput"/>
 			<channel id="activeInputShortName" typeId="activeInputShortName"/>
 			<channel id="activeInputLongName" typeId="activeInputLongName"/>
@@ -68,36 +67,29 @@
 	<!-- Channel types -->
 	<channel-type id="volumeDB" advanced="true">
 		<item-type>Number</item-type>
-		<label>@text/channel-type.anthem.volumedb.label</label>
-		<description>@text/channel-type.anthem.volumedb.description</description>
+		<label>Volume dB</label>
+		<description>Set the volume level dB between -90 and 0</description>
 		<category>SoundVolume</category>
 		<state min="-90" max="0" step="1" pattern="%.0f dB"/>
 	</channel-type>
 
-	<channel-type id="muteToggle" advanced="true">
-		<item-type>Switch</item-type>
-		<label>@text/channel-type.anthem.mutetoggle.label</label>
-		<description>@text/channel-type.anthem.mutetoggle.description</description>
-		<category>Switch</category>
-	</channel-type>
-
 	<channel-type id="activeInput">
 		<item-type>Number</item-type>
-		<label>@text/channel-type.anthem.activeinput.label</label>
-		<description>@text/channel-type.anthem.activeinput.description</description>
+		<label>Active Input</label>
+		<description>Selects the active input source</description>
 	</channel-type>
 
 	<channel-type id="activeInputShortName">
 		<item-type>String</item-type>
-		<label>@text/channel-type.anthem.activeinputshortname.label</label>
-		<description>@text/channel-type.anthem.activeinputshortname.description</description>
+		<label>Active Input Short Name</label>
+		<description>Short friendly name of the active input source</description>
 		<state readOnly="true"></state>
 	</channel-type>
 
 	<channel-type id="activeInputLongName" advanced="true">
 		<item-type>String</item-type>
-		<label>@text/channel-type.anthem.activeinputlongname.label</label>
-		<description>@text/channel-type.anthem.longname.description</description>
+		<label>Active Input Long Name</label>
+		<description>Long friendly name of the active input source</description>
 		<state readOnly="true"></state>
 	</channel-type>
